split type inference into its own module
